#588 Fix: Renamed files to remove 'global'
